DMP-847 share data between steps
