+++ src/main/java/uk/gov/hmcts/darts/automation/utils/EventApi.java
@@ -1,0 +1,38 @@
+package uk.gov.hmcts.darts.automation.utils;
+
+import java.sql.Connection;
+import java.sql.DriverManager;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
+import java.sql.SQLException;
+import java.sql.Statement;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+import java.time.Duration;
+import java.util.List;
+import java.util.Map;
+
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+
+import org.junit.jupiter.api.Test;
+
+
+public class EventApi extends JsonApi {
+	private static Logger log = LogManager.getLogger("EventApi");
+
+	public EventApi() {
+		super();
+	}
+
+	@Test
+	public void testThis() throws Exception {
+	}
+
+
+
+
+
+}
