+++ src/test/java/uk/gov/hmcts/darts/automation/cucumber/steps/StepDef_api.java
@@ -19,7 +19,7 @@ import uk.gov.hmcts.darts.automation.utils.SharedDriver;
 import uk.gov.hmcts.darts.automation.utils.TestData;
 import uk.gov.hmcts.darts.automation.utils.WaitUtils;
 import uk.gov.hmcts.darts.automation.utils.ReadProperties;
-import uk.gov.hmcts.darts.automation.utils.Api;
+import uk.gov.hmcts.darts.automation.utils.JsonApi;
 import uk.gov.hmcts.darts.automation.utils.JsonUtils;
 
 import org.apache.logging.log4j.LogManager;
@@ -31,12 +31,12 @@ public class StepDef_api extends StepDef_base {
 
 	private static Logger log = LogManager.getLogger("StepDef_api");
 	private static String eventFields = "|message_id|type|sub_type|event_id|courthouse|courtroom|case_numbers|event_text|date_time|case_retention_fixed_policy|case_total_sentence|";
-	private Api api;
+	private JsonApi jsonApi;
 	
 	
 	public StepDef_api(SharedDriver driver, TestData testdata, NavigationShared ns) {
 		super(driver, testdata, ns);
-		api = new Api();
+		jsonApi = new JsonApi();
 	}
 	
 	String getValue(Map<String, String> map, String column) {
@@ -77,17 +77,17 @@ public class StepDef_api extends StepDef_base {
 				testdata.getProperty("date_time"),
 				testdata.getProperty("case_retention_fixed_policy"),
 				testdata.getProperty("case_total_sentence"));
-		api.postApi("events", json);
+		jsonApi.postApi("events", json);
 	}
 	
 	@When("I call POST {word} API using json body:")
 	public void callPostApiWithJsonBody(String endPoint, String docString) throws Exception {
-		api.postApi(endPoint, docString);
+		jsonApi.postApi(endPoint, docString);
 	}
 	
 	@When("I call GET {word} API")
 	public void callGetApiWithJsonBody(String endPoint) throws Exception {
-		api.getApi(endPoint);
+		jsonApi.getApi(endPoint);
 	}
 
 // sample cucumber:
@@ -99,7 +99,7 @@ public class StepDef_api extends StepDef_base {
 	public void callGetCaseHearingsApi(List<Map<String, String>> dataTable) throws Exception {
 		for (Map<String, String> map : dataTable) {
 			map.get("case_number");
-			api.getApi("cases/search", map);
+			jsonApi.getApiWithQueryParams("cases/search", map);
 		}
 	}
