+++ src/test/java/uk/gov/hmcts/darts/automation/cucumber/testRunner/TestRunner_Cucumber.java
@@ -24,24 +24,21 @@ import static io.cucumber.junit.platform.engine.Constants.EXECUTION_DRY_RUN_PROP
 
 @Suite
 @IncludeEngines("cucumber")
-@IncludeTags(value = "smoketest")
+@IncludeTags(value = "API_TEST")
 //@IncludeTags(value = "smoketest | DMP-407")
 @ExcludeTags(value = "broken")
 @ConfigurationParameter(key = PLUGIN_PROPERTY_NAME, value = "pretty")
-//@ConfigurationParameter(key = PLUGIN_PROPERTY_NAME, value = "html:target/html/results.html")
 @ConfigurationParameter(key = PLUGIN_PROPERTY_NAME, value = "html:reports/html/results.html, junit:reports/junit/cucumber.xml")
 @ConfigurationParameter(key = GLUE_PROPERTY_NAME, value = "uk/gov/hmcts/darts/automation/cucumber/steps")
 @ConfigurationParameter(key = SNIPPET_TYPE_PROPERTY_NAME, value = "camelcase")
 @ConfigurationParameter(key = EXECUTION_DRY_RUN_PROPERTY_NAME, value = "false")
 @SelectClasspathResource(value = "cucumber/features")
 
-@ConfigurationParameter(key = PARALLEL_EXECUTION_ENABLED_PROPERTY_NAME, value = "false")
-//@ConfigurationParameter(key = PARALLEL_CONFIG_STRATEGY_PROPERTY_NAME, value = "dynamic")
+@ConfigurationParameter(key = PARALLEL_EXECUTION_ENABLED_PROPERTY_NAME, value = "false")	// false or dynamic
 @ConfigurationParameter(key = PARALLEL_CONFIG_STRATEGY_PROPERTY_NAME, value = "fixed")
 @ConfigurationParameter(key = PARALLEL_CONFIG_DYNAMIC_FACTOR_PROPERTY_NAME, value = "8")
 @ConfigurationParameter(key = PARALLEL_CONFIG_FIXED_PARALLELISM_PROPERTY_NAME, value = "8")
-//@ConfigurationParameter(key = EXECUTION_MODE_FEATURE_PROPERTY_NAME, value = "same_thread")
-@ConfigurationParameter(key = EXECUTION_MODE_FEATURE_PROPERTY_NAME, value = "concurrent")
+@ConfigurationParameter(key = EXECUTION_MODE_FEATURE_PROPERTY_NAME, value = "concurrent")	// same_thread or concurrent
 
 
 public class TestRunner_Cucumber {
